requirement.txt
Flask
flask-cors
tensorflow
numpy
Pillow
opencv-python
gunicorn
h5py
